+++ walking-beans-backend/src/main/resources/mappers/menuMapper.xml
@@ -2,5 +2,10 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="walking_beans.walking_beans_backend.mapper.MenuMapper">
+            <select id="">
+                SELECT *
+                FROM MENU
+                WHERE store_id = #{storeid}
 
+            </select>
 </mapper>
