+++ walking-beans-backend/src/main/resources/mappers/storeMapper.xml
@@ -3,4 +3,10 @@
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="walking_beans.walking_beans_backend.mapper.StoreMapper">
 
+    <select id="">
+        SELECT *
+        FROM store
+
+    </select>
+
 </mapper>
